Move progress items to ...core.ui.progress
