+++ feature-odyssey/src/main/kotlin/com/hedvig/android/odyssey/resolution/ui/ResolutionRoot.kt
@@ -8,8 +8,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import com.hedvig.android.core.ui.FullScreenProgressOverlay
 import com.hedvig.android.core.ui.appbar.TopAppBarWithClose
+import com.hedvig.android.core.ui.progress.FullScreenProgressOverlay
 import com.hedvig.android.odyssey.model.Resolution
 import com.hedvig.android.odyssey.resolution.ResolutionViewModel
 import com.hedvig.android.odyssey.resolution.ResolutionViewState
